+++ tools/freezed_to_ts/test_data/address.dart
@@ -6,8 +6,6 @@ import 'package:flutter/foundation.dart';
 part 'address.freezed.dart';
 part 'address.g.dart';
 
-enum AddressType { home, work, other }
-
 @freezed
 class Address with _$Address {
   const factory Address({