+++ packages/anyhoo_logging/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Add Sentry logging
+
 ## 0.0.1
  
 * Add logging package